Responsive Typography
Calculated and set :root variables in JavaScript. 
As development browser (Brave) does not recognize CSS clamp() function, I created a custom _clamp() function in JavaScript to return values for CSS :root variables that will control font-sizes and spacings.
This function is called on initialization, as well as by a "resize" eventListener.