correction: eventListener conflict
The conflict between eventListeners on the horizontal timeline and eventListeners for the Projects slide-out panel, was tracked via a series of tests removing subsequent functions such as smoothScrool(), navbarHide() and other features.
The problem was traced to the initGlitch() function, and specifically to code lines related to an intersectionObserver. As the glitch was not using an observer, these lines were removed.
Testing revealed that the Projects panel and Employment timeline now operate correctly.